+++ simple-app/src/store/messages/actions.js
@@ -1,3 +1,6 @@
+import { getCurrentDate } from "../../utils/dateUtils";
+import { AUTHORS } from '../../utils/constants'
+
 export const ADD_MESSAGE = "MESSAGES::ADD_MESSAGE";
 export const DELETE_MESSAGE = "MESSAGES::DELETE_MESSAGE";
 
@@ -19,3 +22,18 @@ export const deleteMessage = (chatId, id) => ({
         id
     },
 });
+
+let timeout;
+/*
+//redux-thunk example
+
+export const addMessageWithReply = (chatId, text, author, date) => (dispatch) => {
+    dispatch(addMessage(chatId, text, author, date));
+    if (author === AUTHORS.HUMAN) {
+        clearTimeout(timeout);
+        timeout = setTimeout(() => {
+            dispatch(addMessage(chatId, 'I\'m a bot', AUTHORS.BOT, getCurrentDate()));
+        }, 1000);
+    }
+}
+*/
